feat(compat): extract compatibility data to oxc_compat crate (#13932)
Extracted compatibility data to a new `oxc_compat` crate so that we can use this information from the minifier as well.
