+++ crates/oxc_compat/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "oxc_compat"
+version = "0.90.0"
+authors.workspace = true
+categories.workspace = true
+edition.workspace = true
+homepage.workspace = true
+include = ["/src"]
+keywords.workspace = true
+license.workspace = true
+publish = true
+repository.workspace = true
+rust-version.workspace = true
+description.workspace = true
+
+[lints]
+workspace = true
+
+[lib]
+test = true
+doctest = false
+
+[dependencies]
+cow-utils = { workspace = true }
+oxc-browserslist = { workspace = true }
+oxc_syntax = { workspace = true }
+rustc-hash = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
