+++ crates/oxc_compat/README.md
@@ -1,0 +1,1 @@
+# Oxc Compat
