+++ crates/oxc_compat/src/es_target.rs
