+++ crates/oxc_transformer/Cargo.toml
@@ -21,10 +21,10 @@ test = true
 doctest = false
 
 [dependencies]
-oxc-browserslist = { workspace = true }
 oxc_allocator = { workspace = true }
 oxc_ast = { workspace = true }
 oxc_ast_visit = { workspace = true }
+oxc_compat = { workspace = true }
 oxc_data_structures = { workspace = true, features = ["assert_unchecked", "inline_string", "rope", "slice_iter", "stack"] }
 oxc_diagnostics = { workspace = true }
 oxc_ecmascript = { workspace = true }
@@ -37,7 +37,6 @@ oxc_traverse = { workspace = true }
 
 base64 = { workspace = true }
 compact_str = { workspace = true }
-cow-utils = { workspace = true }
 indexmap = { workspace = true }
 itoa = { workspace = true }
 memchr = { workspace = true }
